Portability fix for printing pid
We cannot rely on "%d" to print the pid correctly; POSIX specifies that:
    pid_t... shall be signed integer types.
    https://pubs.opengroup.org/onlinepubs/9699919799/basedefs/sys_types.h.html

That page also says:
    The implementation shall support one or more programming
    environments in which the widths of ... pid_t, ...  are no greater
    than the width of type long.
but if we were to rely on that, we'd need to check that we were in the
relevant programming environment.  Easier just to use "%jd".

+++ tests/events/main.c
@@ -1,3 +1,4 @@
+#include <stdint.h>
 #include <stdio.h>
 #include <stdlib.h>
 #include <unistd.h>
@@ -143,7 +144,7 @@ write_pidfile(const char * filename)
 		warnp("fopen(%s)", filename);
 		goto err0;
 	}
-	if (fprintf(fp, "%d", getpid()) < 0) {
+	if (fprintf(fp, "%jd", (intmax_t)getpid()) < 0) {
 		warnp("fprintf");
 		goto err1;
 	}
